moving default storage to a node folder. Combining custom-utils files
