+++ lib/persistence.js
@@ -27,7 +27,7 @@ function Persistence (options) {
   this.corruptAlertThreshold = options.corruptAlertThreshold !== undefined ? options.corruptAlertThreshold : 0.1;
 
   if(!this.inMemoryOnly)
-    this.storage = options.storage || require('./storage');
+    this.storage = options.storage || require('../nodejs/storage');
 
   if (!this.inMemoryOnly && this.filename && this.filename.charAt(this.filename.length - 1) === '~') {
     throw new Error("The datafile name can't end with a ~, which is reserved for crash safe backup files");
