+++ lib/tree/avl.js
@@ -2,7 +2,7 @@
  * Self-balancing binary search tree using the AVL implementation
  */
 var BinarySearchTree = require('./bst')
-  , customUtils = require('./custom-utils')
+  , customUtils = require('../custom-utils')
   , _ = require('lodash')
   ;
