+++ lib/tree/bst.js
@@ -1,7 +1,7 @@
 /**
  * Simple binary search tree
  */
-var customUtils = require('./custom-utils');
+var customUtils = require('../custom-utils');
 
 
 /**
