--- react-native/path.js
@@ -1,41 +1,0 @@
-/*
- * Polyfill for the core Node.js path module
- */
-
-function assertPath(path) {
-  if (typeof path !== 'string') {
-    throw new TypeError('Path must be a string. Received ' + path);
-  }
-}
-
-module.exports = {
-  // This only supports Unix filepaths which both Android and iOS use
-  // Taken from the core Node.js path module.
-  dirname: function(path) {
-    assertPath(path);
-    if (path.length === 0)
-      return '.';
-    var code = path.charCodeAt(0);
-    var hasRoot = (code === 47/*/*/);
-    var end = -1;
-    var matchedSlash = true;
-    for (var i = path.length - 1; i >= 1; --i) {
-      code = path.charCodeAt(i);
-      if (code === 47/*/*/) {
-        if (!matchedSlash) {
-          end = i;
-          break;
-        }
-      } else {
-        // We saw the first non-path separator
-        matchedSlash = false;
-      }
-    }
-
-    if (end === -1)
-      return hasRoot ? '/' : '.';
-    if (hasRoot && end === 1)
-      return '//';
-    return path.slice(0, end);
-  }
-};
